Fix exceeds the access limited of the memory space when shared memory
Detial:
        * The page fault handler access pages by the memory and MSHRs constrain, when shared memory between applications, the application ID should replace into -1
        * Revise some data format to avoid the overflow

Detial:
        *Some MMU and MemoryController API use int to store the numOfByte, seems overflow

+++ src/MemoryController.cpp
@@ -27,7 +27,7 @@ MemoryController::MemoryController(unsigned long long storage_limit, int page_si
     // storageLimit = 0;
     // for (auto storage : storages) storageLimit += storage.second->storageSize;
 
-    for (int i = 0; i < PRE_ALLOCATE_SIZE / PAGE_SIZE; i++)
+    for (unsigned long long i = 0; i < PRE_ALLOCATE_SIZE / PAGE_SIZE; i++)
     {
         createPage();
     }
@@ -122,13 +122,13 @@ MemoryController::createPage()
  * ================================================================================================
  */
 Page*
-MemoryController::memoryAllocate (int numOfByte)
+MemoryController::memoryAllocate (unsigned long long numOfByte)
 {
     ASSERT(numOfByte != 0, "Try to allocate memory to empty data");
 
     Page* headPage;
     Page* prevPage;
-    for (int i = 0; i < ceil((double)numOfByte / PAGE_SIZE); i++)
+    for (unsigned long long i = 0; i < ceil((double)numOfByte / PAGE_SIZE); i++)
     {
         if (availablePageList.empty()) createPage();
